chore: create favorites file
